ai_service/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.17
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
pytest==8.3.3
httpx==0.27.2
